// src/app/api/models/lista-electoral.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Jurisdiccion } from '../models/jurisdiccion';
export interface ListaElectoral {
  activo?: boolean;
  idListaElectoral?: string;
  jurisdiccion?: Jurisdiccion;
  nombreLista?: string;
  numeroLista?: number;
  siglas?: string | null;
}